WiP for sendmail log parser and redis publisher
- Heavy modifications in Tail module to correctly use FreeBSD KQueue
even in case there are too many files in a directory. The problem is
that, when trying to kqueue an entire directory with many files, the
file descriptors are too many, so if there are lots of events the
load on our application is heavy, even when we are interested in too
few kinds of events (say a couple of files in /var/log). So, the
Tail module monitors the specified files *only* and not the directory
than contains them. This eases the load. However, when a file gets
deleted, there is no create event when the same filename is created.
To work around this, when and only when some files cannot be opened,
we setup a timer (default 5 seconds) to retry and setup the KQueue.
When we have successfully opened all the files, we no longer need that
retry timer and we get rid of it. If a file gets deleted at some
point in the future, we resume our retry scheme until we can open
all the files again. This should make the Tail module light enough.
- Initial version of a Sendmail log parser. We are only interested in
from= events, i.e. log entries that signify an incoming mail. Many more
patterns will probably be added in the future.
- Initial version of a Redis publisher. The intention is to publish
the events to a Redis PubSub and have some logstash listen to this
PubSub channel.
- Better logging, more functionality to the base module Role.
- Some minor modifications to the main program.
- A sample conf file demonstrating a typical usage scenario.
- A .gitignore to unclutter the git status output.