Fix Tail module to detect file rotation
AnyEvent::FileSys::Notify with KQueue backend will erroneously
report a file modification when the file has been rotated by
newsyslog. The IO::KQueue correctly reports the deletion,
but A:F:N does not detect it. This is because newsyslog will
instantaneously replace the file, so, looking at the filesystem
will not show that something is missing. Unfortunately A:F:N
does not use the bit fields returned by IO::KQueue, but just sees
the event and tries to figure out what happened on its own. So,
when a file is instantaneously replaced (newsyslog in FreeBSD for
example), the A:F:N will not spot the change. So, we are checking
ourselves what happened, e.g. by checking the inode number, etc.
Tests seem to indicate that with these fixes the Tail module now
is able to detect rotations correctly.