Add SIGHUP handling to daemon
Daemon process can now receive SIGHUP and try to reload the
configuration if the conf file has been modified. To achieve this
robustly, we handle the signal and inside the signal handler we
"send" an AnyEvent Condition Variable. So, all the actual work
will happen *outside* the signal handler, which is probably the
most safe way to do it. We take care to rearm the condition variable
when we finish, so that subsequent invocations of the signal handler
will also be able to "send" an armed condvar.